Create an convertQGISDateTimeFormatToMoment utils method to convert Qgis date time format to Moment date time format
