+++ src/app/gui/inputs/datetimepicker/service.js
@@ -1,4 +1,4 @@
-const {inherit, base} = require('core/utils/utils');
+const {inherit, base, convertQGISDateTimeFormatToMoment} = require('core/utils/utils');
 const ApplicationService = require('core/applicationservice');
 const Service = require('gui/inputs/service');
 
@@ -17,12 +17,7 @@ proto.getLocale = function() {
 };
 
 proto.convertQGISDateTimeFormatToMoment = function(datetimeformat) {
-  datetimeformat = datetimeformat.replace('yyyy', 'YYYY');
-  const matchDayInDate = datetimeformat.match(/d/g);
-  if (matchDayInDate && matchDayInDate.length < 3) {
-    datetimeformat = datetimeformat.replace('d'.repeat(matchDayInDate.length), 'D'.repeat(matchDayInDate.length))
-  }
-  return datetimeformat
+  return convertQGISDateTimeFormatToMoment(datetimeformat);
 };
 
 proto.setValidatorOptions = function(options) {
